+++ frontend/simple/main.js
@@ -9,7 +9,7 @@ import './views/utils/translations.js'
 import './views/utils/transitions.js'
 import router from './controller/router.js'
 import * as db from './model/database.js'
-import Sidebar from './views/containers/Sidebar/Sidebar.vue'
+import Sidebar from './views/containers/sidebar/Sidebar.vue'
 import store from './model/state.js'
 import { LOGOUT } from './utils/events'
 import Modal from './views/components/Modal/Modal.vue'
