+++ frontend/simple/views/containers/sidebar/Activity.vue
