+++ frontend/simple/views/containers/sidebar/GroupsList.vue
