+++ frontend/simple/views/containers/sidebar/Sidebar.vue
